+++ src/components/signal-detail-sidebar/index.js
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext, useEffect, useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 
 import {
@@ -6,7 +6,6 @@ import {
   CardBody,
   HeadingText,
   InlineMessage,
-  Link,
   NerdGraphQuery,
   SectionMessage,
   navigation,
@@ -19,7 +18,7 @@ import {
   workloadEntitiesQuery,
   workloadEntitiesViolationsFromGuidsArray,
 } from '../../queries';
-import { formatTimestamp, isWorkload } from '../../utils';
+import { formatTimestamp, getOrigin, isWorkload } from '../../utils';
 import { MAX_PARAMS_IN_QUERY, SIGNAL_TYPES, STATUSES } from '../../constants';
 
 import typesList from '../../../nerdlets/signal-selection/types.json';
@@ -163,6 +162,19 @@ const SignalDetailSidebar = ({ guid, name, type, data, timeWindow }) => {
     loadWorkloadIncidents(guid);
   }, [guid, type, data, timeWindow]);
 
+  const signalLink = useMemo(() => {
+    if (!hasAccessToEntity || !guid) return null;
+    const { pathname, search } = navigation.getOpenEntityLocation(guid) || {};
+    if (!pathname) return null;
+
+    const link = `${getOrigin()}${pathname}${search || ''}`;
+    return (
+      <a className="detail-link" href={link} target="_blank" rel="noreferrer">
+        {detailLinkText}
+      </a>
+    );
+  }, [detailLinkText, hasAccessToEntity, guid]);
+
   return (
     <div className="signal-detail-sidebar">
       <div className="signal-header">
@@ -179,15 +191,7 @@ const SignalDetailSidebar = ({ guid, name, type, data, timeWindow }) => {
                   : 'Alert condition'
               }`}
             </HeadingText>
-            {hasAccessToEntity ? (
-              <Link
-                className="detail-link"
-                to={navigation.getOpenEntityLocation(guid)}
-                onClick={(e) => e.target.setAttribute('target', '_blank')}
-              >
-                {detailLinkText}
-              </Link>
-            ) : null}
+            {signalLink}
             {timeWindow ? (
               <InlineMessage
                 className="detail-time-info"