+++ src/constants/entities.js
@@ -2,3 +2,6 @@ export const WORKLOAD = {
   DOMAIN: 'NR1',
   TYPE: 'WORKLOAD',
 };
+
+export const SKIP_ENTITY_TYPES_NRQL =
+  "domain NOT IN ('AIOPS', 'PROTO', 'REF', 'VIZ')";